+++ migrations/121_pl_courses__commit_hash__add.sql
@@ -1,0 +1,3 @@
+-- Someday git will move away from SHA-1 hashes, so we won't constrain this
+-- column to 40 characters.
+ALTER TABLE pl_courses ADD COLUMN IF NOT EXISTS commit_hash TEXT;